Extracted pipeline installer from the pipeline wrapper
